third_party/WebKit/LayoutTests/editing/selection/user-select-js-property-expected.txt

PASS selection.type is "Range"
PASS selection.type is "None"
PASS successfullyParsed is true

TEST COMPLETE
